requirements.txt
altair
numpy
pandas
Pillow
matplotlib
scipy
pydeck
streamlit